+++ nitro/__init__.py
@@ -21,6 +21,6 @@
 
 from nitro.app import Nitro
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
 __all__ = ["Nitro", "__version__"]